Add uncertainty to InteractiveDataDisplay.Polyline. Remove UncertainLinePlot from chart viewer
