+++ src/viewer/Band.ts
@@ -1,7 +1,6 @@
 ﻿/// <reference path="../../typings/jquery/jquery.d.ts" />
 /// <reference path="Utils.ts" />
 /// <reference path="PlotRegistry.ts" />
-/// <reference path="UncertainLinePlot.ts" />
 declare var InteractiveDataDisplay: any;
 declare var Microsoft: any;
