+++ src/viewer/Utils.ts
@@ -170,37 +170,6 @@ module ChartViewer {
         return max;
     }
 
-    export function CutArray(arr: any[], len: number) {
-        if (arr.length > len) {
-            var result = new Array(len);
-            for (var i = 0; i < len; i++) {
-                result[i] = arr[i];
-            }
-            return result;
-        } else {
-            return arr;
-        }
-    }
-
-    export function IsOrderedArray(arr: any[]) {
-        if (arr.length <= 1)
-            return true;
-        else {
-            if (isNaN(arr[1]))
-                return false;
-            if (isNaN(arr[2]))
-                return false;
-
-            var diff = arr[1] - arr[0];
-            for (var i = 2; i < arr.length; i++) {
-                var diff_i = arr[i] - arr[i - 1];
-                if (diff * diff_i < 0)
-                    return false;
-            }
-            return true;
-        }
-    }
-
     export function getFormatter(arr, getRange) {
         var range = getRange(arr);
         var formatter = MathUtils.getPrintFormat(range.min, range.max, (range.max - range.min) / 4);
