+++ test/manual/Input data to idd and chartViewer (line).html
@@ -1,0 +1,126 @@
+﻿<!DOCTYPE html>
+<html xmlns="http://www.w3.org/1999/xhtml">
+<head>
+    <meta http-equiv="X-UA-Compatible" content="IE=10; IE=Edge" />
+    <title>Input data to idd and chart viewer</title>
+    <link rel="stylesheet" type="text/css" href="../../dist/chartViewer.css" />
+    <link rel="stylesheet" type="text/css" href="../../dist/idd.css" />
+    <link rel="stylesheet" type="text/css" href="../../src/css/IDDTheme.css" />
+
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/modernizr/2.8.3/modernizr.min.js"></script>
+    <script src="https://code.jquery.com/jquery-2.1.4.min.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/rxjs/3.1.2/rx.lite.min.js"></script>
+    <script src="https://code.jquery.com/ui/1.12.0-beta.1/jquery-ui.min.js"></script>
+    <script type="text/javascript" src="../../dist/idd.js"></script>
+    <script type="text/javascript" src="../../dist/chartViewer.js"></script>
+
+    <script type="text/javascript">
+        $(document).ready(function () {
+            var dataSet = {
+                "line only with y": {
+                    y: [1, -0.2, 0, -0.3, -0.5]
+                },
+                "line": {
+                    y: [1, -0.2, 0, -0.3, -0.5],
+                    x: [-0.5, 0.5, 0.6, 0.7, 0.9],
+                },
+                "red line": {
+                    y: [1, -0.2, 0, -0.3, -0.5],
+                    x: [-0.5, 0.5, 0.6, 0.7, 0.9],
+                    thickness: 10,
+                    stroke: 'red'
+                },
+                "line with lineCap and lineJoin": {
+                    y: [1, -0.2, 0, -0.3, -0.5],
+                    x: [-0.5, 0.5, 0.6, 0.7, 0.9],
+                    thickness: 10,
+                    lineCap: 'round',
+                    lineJoin: 'bevel'
+                },
+                "line with uncertainty (u68, l68)": {
+                    x: [1, 2, 3, 4, 5],
+                    y: {
+                        median: [0.003, 0.1, 3, 4, 5],
+                        lower68: [0.001, 0.05, 2.5, 3.5, 4.5],
+                        upper68: [0.004, 0.2, 3.7, 4.8, 5.5]
+                    },
+                    fill68: 'darkgreen',
+                },
+                "line wit uncertainty (u95, l95)": {
+                    x: [1, 2, 3, 4, 5],
+                    y: {
+                        median: [0.003, 0.1, 3, 4, 5],
+                        lower95: [0.001, 0.01, 2, 3, 4],
+                        upper95: [0.005, 0.2, 4, 5, 6]
+                    },
+                    fill95: 'green'
+                },
+                "line with uncertainty": {
+                    x: [1, 2, 3, 4, 5],
+                    y: {
+                        median: [0.003, 0.1, 3, 4, 5],
+                        lower95: [0.001, 0.01, 2, 3, 4],
+                        upper95: [0.005, 0.2, 4, 5, 6],
+                        lower68: [0.001, 0.05, 2.5, 3.5, 4.5],
+                        upper68: [0.004, 0.2, 3.7, 4.8, 5.5]
+                    },
+                },
+                "line with treatAs = 'trajectory'": {
+                    y: [0, -0.2, -0.5, -0.3, 1],
+                    x: [0.6, 0.5, 0.9, 0.7, -0.5],
+                    thickness: 4,
+                    stroke: 'orange',
+                    treatAs: 'trajectory'
+                },
+                "line with treatAs = 'function'": {
+                    y: [0, -0.2, -0.5, -0.3, 1],
+                    x: [0.6, 0.5, 0.9, 0.7, -0.5],
+                    thickness: 4,
+                    stroke: 'orange',
+                    treatAs: 'function'
+                },
+                "line with uncertainty and treatAs = 'trajectory'": {
+                    x: [3, 2, 5, 1, 4],
+                    y: {
+                        median: [0.003, 0.1, 3, 4, 5],
+                        lower95: [0.001, 0.01, 2, 3, 4],
+                        upper95: [0.005, 0.2, 4, 5, 6],
+                        lower68: [0.001, 0.05, 2.5, 3.5, 4.5],
+                        upper68: [0.004, 0.2, 3.7, 4.8, 5.5]
+                    },
+                    treatAs: 'trajectory'
+                },
+                "line with uncertainty and traetAs = 'function'": {
+                    x: [3, 2, 5, 1, 4],
+                    y: {
+                        median: [0.003, 0.1, 3, 4, 5],
+                        lower95: [0.001, 0.01, 2, 3, 4],
+                        upper95: [0.005, 0.2, 4, 5, 6],
+                        lower68: [0.001, 0.05, 2.5, 3.5, 4.5],
+                        upper68: [0.004, 0.2, 3.7, 4.8, 5.5]
+                    },
+                    treatAs: 'function'
+                },
+            };
+            
+            for (var data in dataSet) {
+                var div = $('<div style="width:1700px; height: 600px">').appendTo(document.body);
+                var title = $('<h2>' + data + '</h2>').appendTo(div);
+                var chart = $('<div data-idd-plot="chart" style="width: 600px; height: 500px; float: left;"></div>').appendTo(div);
+                var plot = InteractiveDataDisplay.asPlot(chart);
+                var markers = plot.polyline("line", dataSet[data]);
+
+                var chartViewer = $('<div style="width: 1000px; height: 500px; float: right;"></div>').appendTo(div);
+                ChartViewer.show(chartViewer, {
+                    "line": Plot.line(dataSet[data]),
+                });
+            }
+        });
+    </script>
+</head>
+<body>
+    <header>
+        <h1 align="center">Input data to idd and chart viewer</h1>
+    </header>
+</body>
+</html>
